requirements.txt
beautifulsoup4
lxml
PyYAML
sentencepiece
tensorboard
torch
tqdm
